requirements.txt
kafka-python==2.0.2
apache-flink==1.16.0
python-dotenv==1.0.0